Add a string representation of initializer expressions to summaries.
This is preparatory work for #35418.  The string representation is
produced, and tested via unit tests, but not yet used.  In follow-up
CLs we will begin modifying the analyzer to make use of the string
representation, guarded by a flag at first to avoid incompatibilities
with older summary files.  When it is safe to do so we'll hardcode the
flag and remove the old RPN representation.

Change-Id: Iab92ff742930f39761a9df6630f199359d76fcf6
Reviewed-on: https://dart-review.googlesource.com/c/88231
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>