Temporary desugaring of set literals.
When compiling for backends that have not yet implemented support for
the set literal kernel node, transform all such literals into either a
Set() creation with a sequence of add() calls (for non-const literals)
or a const map literal wrapped in a const _UnmodifiableSet object
(for const literals).

The transformation is only performed on bodies or field initializers
that contain at least one actual set literal (i.e. not including
ambiguous set/map literals that were resolved as map literals), so as
to minimize the overhead for code not containing any set literals.

Change-Id: I80318717f0e8481a73f7034e18beea6494fd44bc
Reviewed-on: https://dart-review.googlesource.com/c/87326
Commit-Queue: Aske Simon Christensen <askesc@google.com>
Reviewed-by: Peter von der Ahé <ahe@google.com>