Don't catch call to buildOutlines
This was meant to be a bandage until a real fix landed
(which it did a few days after the 'fix' landed).
While it might still 'fix' something, it also does so silently meaning
that we won't ever catch it if we keep this try block.

Change-Id: I12b69ca06accd0b3fd3c51e2c292e61157c40f71
Reviewed-on: https://dart-review.googlesource.com/c/88571
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Peter von der Ahé <ahe@google.com>